organizing LevelUpTab.tsx, it had been getting clogged with too many options, now they are under descriptive menus to help prevent hair from catching fire. also added class, and tab "Item" though implementation is still.... lacking. update will be pushed to Main as well, as the updates are largely organizational, and will feel much nicer to use.
