+++ components/adjustCoreStatSaves.tsx
@@ -3,7 +3,7 @@ import React, {useState} from "react";
 import {useCharacter, useCharacterUpdater} from "@/components/characterUpdater";
 
 
-export default function AdjustCoreStatSaves() {
+export default function AdjustCoreStatSaves(levelUpToolsSkillsStatsCastingHPDisplay: boolean) {
     const character = useCharacter();
     const characterUpdater = useCharacterUpdater();
 
@@ -11,7 +11,7 @@ export default function AdjustCoreStatSaves() {
 
 
     return(
-        <View style={styles.toolBoxStyle}>
+        <View>{levelUpToolsSkillsStatsCastingHPDisplay && <View style={styles.toolBoxStyle}>
             <Pressable onPress={() => {
                 setAdjustCoreStatSavesToolDisplay(!adjustCoreStatSavesToolDisplay);
             }}>
@@ -46,20 +46,20 @@ export default function AdjustCoreStatSaves() {
                     </View>}
                 </Pressable>
                 <Pressable
-                onPress={() => {
-                    characterUpdater({type: "updateCONSaveProf", value: !character.CONSaveProf});
-                }}>
-                {!character.CONSaveProf && <View style={styles.coreStatAdjustButton}>
-                    <Text style={[styles.standard]}>CON Save</Text>
-                    <Text style={[styles.standard]}>not Prof</Text>
-                </View>}
-                {character.CONSaveProf && <View style={[styles.coreStatAdjustButton, {backgroundColor: "green"}]}>
-                    <Text style={[styles.standard]}>CON Save</Text>
-                    <Text style={[styles.standard]}>Proficient</Text>
-                </View>}
-            </Pressable>
+                    onPress={() => {
+                        characterUpdater({type: "updateCONSaveProf", value: !character.CONSaveProf});
+                    }}>
+                    {!character.CONSaveProf && <View style={styles.coreStatAdjustButton}>
+                        <Text style={[styles.standard]}>CON Save</Text>
+                        <Text style={[styles.standard]}>not Prof</Text>
+                    </View>}
+                    {character.CONSaveProf && <View style={[styles.coreStatAdjustButton, {backgroundColor: "green"}]}>
+                        <Text style={[styles.standard]}>CON Save</Text>
+                        <Text style={[styles.standard]}>Proficient</Text>
+                    </View>}
+                </Pressable>
             </View>}
-            {adjustCoreStatSavesToolDisplay && <View style={{flexDirection: "row", alignSelf: "center"}}>
+            {adjustCoreStatSavesToolDisplay && <View style={{flexDirection: "row", alignSelf: "center", marginBottom: 8}}>
                 <Pressable
                     onPress={() => {
                         characterUpdater({type: "updateINTSaveProf", value: !character.INTSaveProf});
@@ -100,7 +100,7 @@ export default function AdjustCoreStatSaves() {
                     </View>}
                 </Pressable>
             </View>}
-        </View>
+        </View>}</View>
     )
 }
